Refactor Command classes into a new package
Previously, all Command classes were in a same package as the Parser
class. While these classes all deal with parsing input, with the growing
number of Command classes, it would be more sensible to put all the
Command classes into a different package.

Hence, this commit refactors them into a different class, while making
the necessary changes to the import statements.
