+++ src/main/java/adam/command/ListCommand.java
@@ -1,12 +1,14 @@
-package adam.parser;
+package adam.command;
+
+import java.util.ArrayList;
 
 import adam.core.TaskList;
 import adam.exceptions.AdamException;
 
 /**
- * Represents a command to exit the program.
+ * Represents a command to list all tasks in the task list.
  */
-public class ByeCommand extends Command {
+public class ListCommand extends Command {
     /**
      * Checks if the input matches the command.
      *
@@ -14,11 +16,11 @@ public class ByeCommand extends Command {
      * @return True if the input matches the command, false otherwise.
      */
     public static boolean isMatch(String input) {
-        return input.equals("bye");
+        return input.equals("list");
     }
 
     /**
-     * Outputs a goodbye message to the user.
+     * Lists all tasks in the task list.
      *
      * @param manager The task list to add the task to.
      * @return The output to show to the user.
@@ -26,16 +28,11 @@ public class ByeCommand extends Command {
      */
     @Override
     public String execute(TaskList manager) throws AdamException {
-        return "Bye. Hope to see you again soon!";
-    }
-
-    /**
-     * Checks if it is an exit command.
-     *
-     * @return True as Bye is an exit command.
-     */
-    @Override
-    public boolean isExit() {
-        return true;
+        ArrayList<String> outputs = manager.listAll();
+        StringBuilder sb = new StringBuilder();
+        for (String output : outputs) {
+            sb.append(output).append("\n");
+        }
+        return sb.toString();
     }
 }
