+++ src/main/java/adam/parser/Parser.java
@@ -4,6 +4,15 @@ import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 
+import adam.command.AddCommand;
+import adam.command.ByeCommand;
+import adam.command.Command;
+import adam.command.DeleteCommand;
+import adam.command.DoneCommand;
+import adam.command.FindCommand;
+import adam.command.ListCommand;
+import adam.command.ListOnCommand;
+import adam.command.UnmarkCommand;
 import adam.exceptions.AdamException;
 import adam.exceptions.InvalidDate;
