+++ src/test/java/adam/parser/TestParser.java
@@ -1,9 +1,15 @@
 package adam.parser;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-
 import org.junit.jupiter.api.Test;
 
+import adam.command.AddCommand;
+import adam.command.ByeCommand;
+import adam.command.DeleteCommand;
+import adam.command.DoneCommand;
+import adam.command.ListCommand;
+import adam.command.ListOnCommand;
+import adam.command.UnmarkCommand;
 import adam.exceptions.AdamException;
 
 /**
